# contracts/002-trader-orders-views/openapi.yaml
openapi: 3.0.3
info:
  title: Money Market Order Processing API
  description: |
    Contract-first HTTP API for Money Market order processing. Baseline intake and lifecycle
    operations from `001-mm-order-processing`; **002-trader-orders-views** adds workspace-scoped
    list surfaces (`/term/*`, `/oncall/*`) so Term and OnCall orders are never mixed on the same
    list response. Workspace **Assigned** lists are **desk-wide** (all assignees) per User Story 2.
    Human-readable mirror: [api-v1.md](./api-v1.md). Trader-facing operations require header
    `X-User-Id` (actor identity; Assigned workspace lists are not filtered by assignee).
  version: 2.0.0
servers:
  - url: /
    description: Relative to deployment base (e.g. http://localhost:8080)
...
paths:
...
  /api/v1/orders:
    post:
      tags: [Intake]
      summary: Receive order (idempotent)
      operationId: receiveOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveOrderRequest"
      responses:
        "201":
          description: New order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiveOrderResponse"
        "200":
          description: Duplicate external reference — existing order returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiveOrderResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/settings/institutions/{institutionCode}/oncall-rates:
    get:
      tags: [OnCallRateSettings]
      summary: List OnCall rate segments for an institution
      operationId: listOnCallRateSegments
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/InstitutionCodePath"
      responses:
        "200":
          description: Segments for the institution (may be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OnCallRateSegmentResponse"
        "404":
          description: Institution not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [OnCallRateSettings]
      summary: Add an OnCall rate to a curve point
      operationId: addOnCallRate
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/InstitutionCodePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddOnCallRateRequest"
      responses:
        "201":
          description: Pending segment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnCallRateSegmentResponse"
        "400":
          description: Validation error (e.g. backdated value date)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Institution not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Curve point already has a pending segment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/settings/institutions/{institutionCode}/oncall-rates/{segmentId}/cancel:
    post:
      tags: [OnCallRateSettings]
      summary: Cancel a pending OnCall rate segment
      operationId: cancelOnCallRate
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/InstitutionCodePath"
        - $ref: "#/components/parameters/SegmentIdPath"
      responses:
        "200":
          description: Segment canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnCallRateSegmentResponse"
        "404":
          description: Segment or institution not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Segment is not pending confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    UserIdHeader:
      name: X-User-Id
      in: header
      required: true
      description: MMXUser identity for Trader-facing and settings endpoints
      schema:
        type: string
        minLength: 1
...
    SegmentIdPath:
      name: segmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InstitutionCodePath:
      name: institutionCode
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 32
...
  schemas:
    ErrorCode:
      type: string
      description: Machine-readable error code
      enum:
        - VALIDATION_ERROR
        - ORDER_NOT_FOUND
        - INVALID_STATUS_TRANSITION
        - UNAUTHORIZED_TRADER
        - DUPLICATE_ORDER
        - INSTITUTION_NOT_FOUND
        - ONCALL_SEGMENT_NOT_FOUND
        - ONCALL_PENDING_EXISTS
        - ONCALL_BACKDATED_VALUE_DATE
        - ONCALL_INVALID_SEGMENT_STATUS
        - ONCALL_SEGMENT_CANCELED
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
...
    OrderType:
      type: string
      enum: [TERM, ON_CALL]
    OrderOperation:
      type: string
      enum: [SUBSCRIPTION, INCREASE, DECREASE, REDEMPTION]
    OrderStatus:
      type: string
      enum: [RECEIVED, ASSIGNED, EXECUTED, ACCOUNTED, CANCELLED, REJECTED]
    Tenor:
      type: string
      description: Term tenor code
      enum: ["1W", "2W", "1M", "3M", "6M", "1Y"]
    NoticePeriod:
      type: string
      enum: ["24H", "48H"]
    ReceiveOrderRequest:
      type: object
      required:
        - externalOrderReference
        - legalEntityCode
        - orderType
        - orderOperation
        - portfolioNumber
        - currency
        - amount
        - valueDate
        - institutionCode
      properties:
        legalEntityCode:
          type: string
          minLength: 3
          maxLength: 3
          description: Owning LegalEntity code (3 characters, e.g. LOC, PAR).
        externalOrderReference:
          type: string
          maxLength: 100
        orderType:
          $ref: "#/components/schemas/OrderType"
        orderOperation:
          $ref: "#/components/schemas/OrderOperation"
        portfolioNumber:
          type: string
          maxLength: 50
        currency:
          type: string
          minLength: 3
          maxLength: 3
        amount:
          type: number
          format: double
          exclusiveMinimum: 0
        valueDate:
          type: string
          format: date
        minimumRate:
          type: number
          format: double
          minimum: 0
          nullable: true
          description: Optional PM execution floor; when omitted, Trader executes subject to best market conditions.
        tenor:
          $ref: "#/components/schemas/Tenor"
        noticePeriod:
          $ref: "#/components/schemas/NoticePeriod"
        sourceContractNumber:
          type: string
          nullable: true
          maxLength: 50
          description: >
            Required when orderOperation is INCREASE, DECREASE, or REDEMPTION (existing MM contract).
            When orderOperation is SUBSCRIPTION, MUST NOT be persisted; if supplied, the service discards it at reception.
        institutionCode:
          type: string
          minLength: 1
          maxLength: 32
          description: Active onboarded institution selected by Portfolio Management at order creation.
    ReceiveOrderResponse:
      type: object
      required: [orderId, status, legalEntityCode]
      properties:
        orderId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/OrderStatus"
        legalEntityCode:
          type: string
          minLength: 3
          maxLength: 3
          description: Owning LegalEntity code persisted on the order.
...
    OnCallRateSegmentStatus:
      type: string
      enum: [PENDING_CONFIRMATION, VALID, CANCELED]
    OnCallNoticePeriod:
      type: string
      enum: [24H, 48H]
    AddOnCallRateRequest:
      type: object
      required: [currency, noticePeriod, rate, valueDate]
      properties:
        currency:
          type: string
          minLength: 1
        noticePeriod:
          $ref: "#/components/schemas/OnCallNoticePeriod"
        rate:
          type: number
          format: double
        valueDate:
          type: string
          format: date
          description: Inclusive segment start; must be today or later.
    OnCallRateSegmentResponse:
      type: object
      required:
        - segmentId
        - institutionCode
        - currency
        - noticePeriod
        - rate
        - valueDate
        - endDate
        - status
      properties:
        segmentId:
          type: string
          format: uuid
        institutionCode:
          type: string
        currency:
          type: string
        noticePeriod:
          $ref: "#/components/schemas/OnCallNoticePeriod"
        rate:
          type: number
          format: double
        valueDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          description: Inclusive end; open segment uses sentinel 2999-12-31.
        status:
          $ref: "#/components/schemas/OnCallRateSegmentStatus"
        validatedAt:
          type: string
          format: date-time
          description: Present when status is VALID.
...